Started to extract usecases to the module
