+++ app/src/main/kotlin/io/github/fobo66/wearmmr/domain/DomainModule.kt
@@ -1,6 +1,5 @@
 package io.github.fobo66.wearmmr.domain
 
-import io.github.fobo66.wearmmr.domain.usecase.RatingComplicationUseCase
 import org.koin.dsl.module
 
 val domainModule = module {
